refactor: Dealer와 Player의 공통 코드 Participant로 분리
